rename terminal_gui dependency
